Add Validation middleware and update constants.js

+++ controllers/job.controller.js
@@ -1,8 +1,11 @@
 import asyncHandler from '../utils/asyncHandler.js'
 import {Job} from '../models/job.model.js'
 import { nanoid } from 'nanoid'
-
-
+import {statusCodes} from '../utils/constants.js'
+import { 
+    NotFoundError,
+    BadRequestError
+ } from '../errors/customErrors.js'
 
 const getAllJobs = asyncHandler(async (req, res)=>{
     const {limit=10, page=1} = req.query
@@ -17,33 +20,36 @@ const getAllJobs = asyncHandler(async (req, res)=>{
         }
     ]
     const jobs = await Job.aggregate(aggregationPipeline)
-    res.status(200).json({jobs})
+    res.status(statusCodes.OK).json({jobs})
 })
 
 const createJob = asyncHandler(async (req, res)=>{
     const {company, position} = req.body
     if (!company || !position){
-        return res.status(400).json({msg: "All feilds are required"})
+        throw new BadRequestError(`All fields are required`) 
     }
    
     const job = await Job.create({company, position})
 
-    res.status(201).json({job})
+    res.status(statusCodes.CREATED).json({job})
 })
 
 const getJob = asyncHandler(async (req, res)=>{
     const {id} = req.params
+    if (!id){
+        throw new BadRequestError("Id is missing")
+    }
     const job = await Job.findById(id)
     if (!job){
-        return res.status(404).json({msg: `No job with id ${id}`})
+        throw new NotFoundError(`No job with id ${id}`)
     }
-    res.status(200).json({job})
+    res.status(statusCodes.OK).json({job})
 })
 
 const updateJob = asyncHandler(async (req, res)=>{
     const {company, position} = req.body
     if (!company && !position){
-        return res.status(400).json("msg: at least one field is required")
+        throw new BadRequestError("At least one field is required")
     }
     const {id} = req.params
     
@@ -53,19 +59,22 @@ const updateJob = asyncHandler(async (req, res)=>{
         {new: true}
     )
     if (!job){
-        return res.status(404).json({msg: `No job with id ${id}`})
+        throw new NotFoundError(`No job with id ${id}`)
     }
-    res.status(200).json({mgs: `Job modified:`, job})
+    res.status(statusCodes.OK).json({mgs: `Job modified:`, job})
 })
 
 const deleteJob = asyncHandler(async (req, res)=>{
     const {id} = req.params
+    if (!id){
+        throw new BadRequestError("Id is missing")
+    }
     const job = await Job.findOneAndDelete({_id: id})
     if (!job){
-        return res.status(404).json({msg: `No job with id ${id}`})
+        throw new NotFoundError(`No job with id ${id}`)
     }
     
-    res.status(200).json({msg: "job deleted"})
+    res.status(statusCodes.OK).json({msg: "job deleted"})
 })
 
 export {
